+++ src/components/common/Card.js
@@ -15,7 +15,7 @@ const Card = (props) => {
   return (
     <CardLayout>
       <h3>{idx}</h3>
-      이름 : {info.username}<br />
+      이름 : {info.name}<br />
       소개 : {info.introduction}<br />
       <h4>학력</h4>
       학력1<br />
