+++ src/containers/SignupContainer.js
@@ -1,24 +1,15 @@
-import React, {useEffect} from 'react';
+import React, { useEffect } from 'react';
 import { useSelector, useDispatch, shallowEqual } from 'react-redux';
 import styled from 'styled-components';
 
 import { addIDPW, addPersonalInfo, addPhone, addAgreement, getSignupStateThunk } from '../modules/reducers/signup';
 
-import Signup01 from '../components/Signup01';
-import Signup02 from '../components/Signup02';
-import Signup03 from '../components/Signup03';
-import Signup04 from '../components/Signup04';
+import Signup01 from '../components/Signup/Signup01';
+import Signup02 from '../components/Signup/Signup02';
+import Signup03 from '../components/Signup/Signup03';
+import Signup04 from '../components/Signup/Signup04';
 
 import * as API from '../lib/api'
-
-const SignupBlock = styled.div`
-  width:100%;
-  
-  .deactivate {
-    display: none;
-  }
-
-`
 
 const SignupContainer = () => {
   const dispatch = useDispatch();
@@ -45,10 +36,15 @@ const SignupContainer = () => {
         // target.classList.add('deactivate');
         // target.parentNode.children[process].classList.remove('deactivate');
         dispatch(addAgreement(formData))
-        dispatch(getSignupStateThunk()).then((data) => {
-          API.postUserData(data)
+        dispatch(getSignupStateThunk()).then(async (data) => {
+          const response = await API.postUserData(data);
+          if (response.status != "ok"){
+            alert("not ok");
+          } else {
+            alert("회원가입이 완료되었습니다");
+            window.location.href = "/signin";
+          }
         })
-        
         break;
       default:
         console.log("onClickNext error");
@@ -64,5 +60,13 @@ const SignupContainer = () => {
     </SignupBlock>
   );
 }
+
+const SignupBlock = styled.div`
+  width:100%;
+  
+  .deactivate {
+    display: none;
+  }
+`;
 
 export default SignupContainer;
